My standard C redefines

+++ include/redefines.h
@@ -4,23 +4,28 @@
 #include <stddef.h>
 #include <stdint.h>
 
+// Unsigned integers
 typedef uint8_t u8;
 typedef uint16_t u16;
 typedef uint32_t u32;
 typedef uint64_t u64;
 
+// Signed integers
 typedef int8_t i8;
 typedef int16_t i16;
 typedef int32_t i32;
 typedef int64_t i64;
 
+// Pointer-sized types
 typedef size_t usize;
-typedef intmax_t isize;
+typedef intptr_t isize;
 
+// Other common aliases
 typedef unsigned long ul;
 typedef unsigned long long ull;
 
+// Floating-poi32 types
 typedef float f32;
 typedef double f64;
 
-#endif
+#endif // REDEFINES_H
